+++ src/main/java/web/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> index();
-    User show(int id);
+    User show(long id);
     void save(User user);
-    void update(int id, User updatedUser);
-    void delete(int id);
+    void update(long id, User updatedUser);
+    void delete(long id);
 }
